util: Move out each category of utility function into separate TU
