+++ src/instr/install.c
